configure: Make PSCI the default boot method
When the boot-wrapper was originally conceived, PSCI was a rather new
feature, so support in contemporary kernels wasn't guaranteed. The
boot-wrapper consequently defaulted to not using PSCI.

Fortunately the times have changed, and most people expect PSCI these
days, so let's enable PSCI by default.

We keep the --enable-psci/--disable-psci configure switch, so it can be
still turned off if needed.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Signed-off-by: Mark Rutland <mark.rutland@arm.com>