configure: Fix default DTB
The DTS files for Arm Ltd. boards and the fastmodel have long been moved
into the arm/ subdirectory of the arm64 kernel tree's DT folder.

Adjust the default path to make this build out of the box.

Also change the default DTB to the more modern FVP RevC model on the
way.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Signed-off-by: Mark Rutland <mark.rutland@arm.com>