Revert "avoid dtc warnings on re-compiling DTB"
As Vladimir Murzin reports in:

  https://lore.kernel.org/linux-arm-kernel/706471d8-a0fd-35fb-4fa0-380bfb1b78e7@arm.com/

Older version of dtc don't support the options to suprress the warnings,
and so where people are stuck with said versions of dtc, using those
options results in a build failure rather than an ignoreable build
warning.

For now, let's revert the warning suppression. In future we can try to
do something smarter.

This reverts commit 1044c77062573985f7c994c3b6cef5695f57e955.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>