Makefile: rework test-dtc-option
Currently we have the test-dtc-option helepr to check whether dtc
supports an option. In subsequent patches we'll want similar for ld, and
it seems likely that we'll want this for other commands in future.

This patch adds a new test-cmd helper that can be used to check whether
a command succeeds (based on its return value), and can be used to
choose between a value if the comand succeeds and a value if the command
fails.

The existing test-dtc-option helper is reworked to use test-cmd. The
helpers are added near the start of the file so that they'll be defined
prior to usage whetever they're used within the file. This is purely for
human comprehension as Make itself does not care about their location
within the Makefile.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Andre Przywara <andre.przywara@arm.com>
Cc: Akos Denke <akos.denke@arm.com>
Cc: Luca Fancellu <luca.fancellu@arm.com>
Cc: Marc Zyngier <maz@kernel.org>
Acked-by: Marc Zyngier <maz@kernel.org>
Link: https://lore.kernel.org/r/20240729143116.1804615-2-mark.rutland@arm.com