aarch64: Start Xen on Armv8-R at EL2
When bootwrapper is compiled with Xen support and it is started
at EL2 on Armv8-R AArch64, keep the current EL and jump to the
Xen image using the SPSR_KERNEL as spsr_el2 value.

The only available boot method on Armv8-R AArch64 when starting
Xen is spin table at the moment, reflect this in the build
configuration.

Signed-off-by: Luca Fancellu <luca.fancellu@arm.com>
[Mark: tweak configure.ac error message]
Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Link: https://lore.kernel.org/r/20241127102300.2822848-1-luca.fancellu@arm.com